+++ example/lib/generate_img_screen.dart
@@ -1,7 +1,6 @@
 import 'dart:async';
 import 'package:chat_gpt_sdk/chat_gpt_sdk.dart';
 import 'package:flutter/material.dart';
-
 import 'constants.dart';
 
 class GenImgScreen extends StatefulWidget {
@@ -20,7 +19,8 @@ class _GenImgScreenState extends State<GenImgScreen> {
   void initState() {
     openAI = OpenAI.instance.build(
         token: token,
-        baseOption: HttpSetup(receiveTimeout: const Duration(seconds: 6)),isLogger: true);
+        baseOption: HttpSetup(receiveTimeout: const Duration(seconds: 6)),
+        isLog: true);
     super.initState();
   }
 
@@ -31,18 +31,16 @@ class _GenImgScreenState extends State<GenImgScreen> {
     super.dispose();
   }
 
-  void _generateImage() async{
+  void _generateImage() async {
     const prompt = "King Snake.";
 
-    final request = GenerateImage(prompt,1);
-    subscription = openAI.generateImageStream(request)
-    .asBroadcastStream()
-    .listen((it) {
+    final request = GenerateImage(prompt, 1,size: ImageSize.size256,response_format: Format.url);
+    subscription =
+        openAI.generateImageStream(request).asBroadcastStream().listen((it) {
       setState(() {
         img = "${it.data?.last?.url}";
       });
     });
-
   }
 
   @override
@@ -56,7 +54,7 @@ class _GenImgScreenState extends State<GenImgScreen> {
                   onPressed: () => _generateImage(),
                   child: const Text("Generate Image"))),
           img == ""
-              ? const SizedBox()
+              ? const Text("Loading...")
               : AspectRatio(
                   aspectRatio: 16 / 9,
                   child: Image.network(img),