+++ lib/src/model/complete_text/response/choices.dart
@@ -2,9 +2,9 @@ class Choices {
   final String text;
   final int index;
   final dynamic logprobs;
-  final String? finish_reason;
+  final String? finishReason;
 
-  Choices(this.text, this.index, this.logprobs, this.finish_reason);
+  Choices(this.text, this.index, this.logprobs, this.finishReason);
 
   factory Choices.fromJson(Map<String, dynamic> json) => Choices(
         json['text'],
@@ -13,12 +13,12 @@ class Choices {
         json['finish_reason'],
       );
 
-  Map<String, dynamic> toJson() => _ChoicesToJson(this);
+  Map<String, dynamic> toJson() => choicesToJson(this);
 
-  Map<String, dynamic> _ChoicesToJson(Choices instance) => <String, dynamic>{
+  Map<String, dynamic> choicesToJson(Choices instance) => <String, dynamic>{
         'text': instance.text,
         'index': instance.index,
         'logprobs': instance.logprobs,
-        'finish_reason': instance.finish_reason,
+        'finish_reason': instance.finishReason,
       };
 }