+++ test/chat_gpt_test.dart
@@ -1,1 +1,5 @@
-void main(){}
+import 'package:chat_gpt_sdk/chat_gpt_sdk.dart';
+
+void main() {
+}
+